+++ backend/src/app/models/review.py
@@ -1,6 +1,6 @@
 import datetime
 
-from sqlalchemy import Column, DateTime, Integer, String
+from sqlalchemy import BigInteger, Column, DateTime, Integer, String
 
 from src.core.database.session import Base
 
@@ -16,7 +16,7 @@ class Review(Base):
     id = Column(Integer, autoincrement=True, primary_key=True, index=True)
     post_id = Column(Integer, nullable=False)
     post_url = Column(String, nullable=False)
-    goods_id = Column(Integer, nullable=False, index=True)
+    goods_id = Column(BigInteger, nullable=False, index=True)
     goods_url = Column(String(length=2048), nullable=True)
     created_at = Column(
         DateTime, default=datetime.datetime.now, index=True,
